Stable. Release in am.
git-svn-id: https://plugins.svn.wordpress.org/contexture-page-security/trunk@324858 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/theme-functions.php
@@ -9,9 +9,9 @@ if(!function_exists('psc_add_user_to_group')){
 /**
  * Can be used by developers to add a user to a group programatically.
  * 
- * @param int $user_id The id of the user to add to a group.
- * @param int $group_id The id of the group to add the user to.
- * @param datetime $expires Unused. If specified, the user wont be able to access restricted content after this date.
+ * @param int $user_id Required. The id of the user to add to a group.
+ * @param int $group_id Required. The id of the group to add the user to.
+ * @param string $expires Optional. A string-formatted datetime (YYYY-MM-DD). If provided, the user wont be able to access group content after this date.
  * @return bool Returns true if user was successfully added to a group.
  */
 function psc_add_user_to_group($user_id,$group_id,$expires=null){
@@ -26,6 +26,10 @@ function psc_add_user_to_group($user_id,$group_id,$expires=null){
     $UserInfo = (int)$wpdb->get_var(
         $wpdb->prepare("SELECT COUNT(*) FROM {$wpdb->users} WHERE {$wpdb->users}.ID = %s",$user_id)
     );
+        
+    //Make sure user isnt already in the group
+    $UserInGroup = $wpdb->prepare('SELECT COUNT(*) FROM `'.$wpdb->prefix.'ps_group_relationships` WHERE grel_group_id=%s AND grel_user_id=%s',$group_id,$user_id);
+    if($wpdb->get_var($sqlUpdateGroup)>0){ return false; }
 
     //If this user doesn't exist
     if($UserInfo === 0){
@@ -49,22 +53,28 @@ function psc_add_user_to_group($user_id,$group_id,$expires=null){
 
 if(!function_exists('psc_update_user_membership')){
 /**
- * INCOMPLETE: Updates expiration date for a user's group membership.
+ * This function is primarily used to update expiration information for a user's group membership, but may be used for other things in future updates.
  * 
- * @param datetime $expires Set datetime to include expiration dat. If null, no expiration.
+ * 
+ * @param int $user_id Required. The users userid. Can usually be retrieved with $current_user->ID.
+ * @param int $group_id Required. The id of the group the user is a member of.
+ * @param string $expires Optional. A date formatted as a string (YYYY-MM-DD). If left blank or null, membership never expires.
  * @return bool Returns true if record was updated. False if query fails.
  */
 function psc_update_user_membership($user_id,$group_id,$expires=null){
     global $wpdb;
     
+    //Fail if the id's aren't numeric
+    if(!is_numeric($user_id) || !is_numeric($group_id)){ return false; }
+    
     //Check if expires date is null
-    $expires = (empty($expires) || strtolower($expires)==='null') ? 'NULL' : "'".$expires."'";
+    $expires = (empty($expires) || strtolower($expires)==='null') ? 'NULL' : "'".date('Y-m-d',strtotime($expires))."'";
     
     $query = sprintf('UPDATE `%1$sps_group_relationships` SET grel_expires=%2$s WHERE `grel_group_id`=\'%3$s\' AND `grel_user_id`=\'%4$s\'',
-            /*1*/$wpdb->prefix,
-            /*2*/$expires,
-            /*3*/$group_id,
-            /*4*/$user_id);
+        /*1*/$wpdb->prefix,
+        /*2*/$expires,
+        /*3*/$group_id,
+        /*4*/$user_id);
     return $wpdb->query($query);
 }
 }
@@ -73,8 +83,8 @@ if(!function_exists('psc_remove_user_from_group')){
 /**
  * Removes a user from a group.
  *
- * @param int $user_id
- * @param int $group_id
+ * @param int $user_id The id of the user to unenroll.
+ * @param int $group_id The group to unenroll from.
  * @return bool Returns true if the query succeeded. False if it failed. 
  */
 function psc_remove_user_from_group($user_id,$group_id){
@@ -83,16 +93,15 @@ function psc_remove_user_from_group($user_id,$group_id){
     if(!is_numeric($user_id) || !is_numeric($group_id)){
         return false;
     }
-    $sqlRemoveUserRel = $wpdb->prepare("DELETE FROM `{$wpdb->prefix}ps_group_relationships` WHERE grel_group_id = '%s' AND grel_user_id = '%s';",
-            $group_id,
-            $user_id);
+    $sqlRemoveUserRel = $wpdb->prepare("DELETE FROM `{$wpdb->prefix}ps_group_relationships` WHERE grel_group_id = %s AND grel_user_id = %s;", $group_id, $user_id);
     return $wpdb->query($sqlRemoveUserRel) == 0;
 }
 }
 
 if(!function_exists('psc_get_groups')){
 /**
- * Gets an assoc array with all the groups.
+ * Gets an assoc array containing a list of all groups. This can also be used to get
+ * a list of only groups belonging to an individual user.
  * 
  * @param int $user_id Optional. Include if you want groups a user is attached to. Leave blank for all groups.
  * @return array Associative array with groups. Format: Group_ID => Group_Title
@@ -102,12 +111,12 @@ function psc_get_groups($user_id=null){
     $array = array();
     
     //Determine if we're looking up groups for a user, or all groups
-    if(is_numeric($userid)){
+    if(is_numeric($user_id)){
         $groups = $wpdb->get_results("
             SELECT * FROM `{$wpdb->prefix}ps_group_relationships`
             JOIN `{$wpdb->prefix}ps_groups`
                 ON {$wpdb->prefix}ps_group_relationships.grel_group_id = {$wpdb->prefix}ps_groups.ID
-            WHERE {$wpdb->prefix}ps_group_relationships.grel_user_id = '{$userid}'
+            WHERE {$wpdb->prefix}ps_group_relationships.grel_user_id = '{$user_id}'
         ");
     }else{
         $groups = $wpdb->get_results("SELECT * FROM `{$wpdb->prefix}ps_groups`");
@@ -115,7 +124,7 @@ function psc_get_groups($user_id=null){
 
     //We only need an ID and a name as a key/value..., so we'll build a new array
     foreach($groups as $group){
-        $array += array($group->ID => $group->group_title);
+        $array[$group->ID] = $group->group_title;
     }
 
     //If multisite is enabled we can better support it...
@@ -145,20 +154,27 @@ if(!function_exists('psc_has_protection')){
  * Recursively checks security for this page/post and it's ancestors. Returns true
  * if any of them are protected or false if none of them are protected.
  * 
- * @param int $post_id The id of the page or post to check.
+ * @param int $post_id Optional. The id of the page or post to check. If left null, will try to check current post id from the loop (if available).
+ * @param bool $dontcheck Optional. Set to true to prevent automatically checking current post id in the loop (if $post_id is null).
  * @return bool If this page or it's ancestors has the "protected page" flag
  */
-function psc_has_protection($post_id){
-    global $wpdb;
+function psc_has_protection($post_id=null,$dontcheck=false){
+    global $wpdb, $post;
+        
+    //If $post_id isnt set, try to get global post id
+    if(empty($post_id) && !$dontcheck && isset($post->ID)){ $post_id=$post->ID; }
     //Fail if the post id isn't numeric
     if(!is_numeric($post_id)){ return false; }
+        
+    //Try to get post meta
+    $mymeta=get_post_meta($post_id,'ctx_ps_security');
     
     //Check permissions for current page
-    if(get_post_meta($post_id,'ctx_ps_security')){
+    if( !empty( $mymeta ) ){
         return true;
     } else {
         //If this isn't protected, lets see if there's a parent...
-        $parent_id = $wpdb->get_var(sprintf('SELECT post_parent FROM $wpdb->posts WHERE `ID` = %s',$post_id));
+        $parent_id = $wpdb->get_var(sprintf('SELECT post_parent FROM %s WHERE `ID` = %s',$wpdb->posts,$post_id));
         //If we have a parent, repeat this check with the parent.
         if ($parent_id != 0)
             return ctx_ps_isprotected_section($parent_id);
